refactor: replace sd-jwt related StorageManager methods with SdJwtUtils in EudiWalletKit and tests
